FOP-2181: Add a test case for the recent fix made in FOP-2174 and XGC-76 regarding source resolution scaling with SVG and images; patch submitted by Robert Meyer
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1448484 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/layoutmgr/inline/ImageLayout.java
@@ -160,8 +160,8 @@ public class ImageLayout implements Constants {
         }
         boolean foundNonAuto = (lenWidth.getEnum() != EN_AUTO || lenHeight.getEnum() != EN_AUTO);
         if (!foundNonAuto) {
-            cwidth = intrinsicSize.width / (sourceResolution / UnitConv.IN2PT);
-            cheight = intrinsicSize.height / (sourceResolution / UnitConv.IN2PT);
+            cwidth = (int)(intrinsicSize.width / ((float)sourceResolution / (float)UnitConv.IN2PT));
+            cheight = (int)(intrinsicSize.height / ((float)sourceResolution / (float)UnitConv.IN2PT));
         }
 
         Dimension constrainedIntrinsicSize;
